changing the layout of the model
