+++ met-web/src/components/engagement/view/EngagementBanner.tsx
@@ -3,7 +3,6 @@ import { ActionContext } from './ActionContext';
 import { Button, Grid, Skeleton } from '@mui/material';
 import { Banner } from '../banner/Banner';
 import { useNavigate } from 'react-router-dom';
-import EmailModal from 'components/survey/submit/EmailModal';
 import { ConditionalComponent } from 'components/common';
 
 export const EngagementBanner = () => {
@@ -20,10 +19,10 @@ export const EngagementBanner = () => {
         <Banner savedEngagement={savedEngagement}>
             <ConditionalComponent condition={!!surveyId && publishedStatus}>
                 <Grid item xs={12} container direction="row" justifyContent="flex-end">
-                    <Button variant="contained" onClick={() => setOpen(true)}>
+                    {/* <Button variant="contained" onClick={() => setOpen(true)}>
                         Share your thoughts
                     </Button>
-                    <EmailModal open={open} handleClose={() => setOpen(false)} />
+                    <EmailModal open={open} handleClose={() => setOpen(false)} /> */}
                 </Grid>
             </ConditionalComponent>
         </Banner>
